requirements.txt
Django==2.0.5
djangorestframework==3.8.2
wheel==0.24.0
